refactor(forms): split oversized form components
